src/third_party/blink/web_tests/external/wpt/html/semantics/scripting-1/the-script-element/module/instantiation-error-3-expected.txt

This is a testharness.js-based test.
FAIL Test that unresolvable cycles lead to SyntaxError events on window and load events on script assert_array_equals: expected property 2 to be "Uncaught SyntaxError: Detected cycle while resolving name 'x' in './cycle-unresolvable-a.js'" but got "Uncaught SyntaxError: Detected cycle while resolving name 'x' in './cycle-unresolvable.js'" (expected array ["Uncaught SyntaxError: Detected cycle while resolving name 'x' in './cycle-unresolvable-a.js'", 1, "Uncaught SyntaxError: Detected cycle while resolving name 'x' in './cycle-unresolvable-a.js'", 2, "Uncaught SyntaxError: Detected cycle while resolving name 'x' in './cycle-unresolvable-a.js'", 3] got ["Uncaught SyntaxError: Detected cycle while resolving name 'x' in './cycle-unresolvable-a.js'", 1, "Uncaught SyntaxError: Detected cycle while resolving name 'x' in './cycle-unresolvable.js'", 2, "Uncaught SyntaxError: Detected cycle while resolving name 'x' in './cycle-unresolvable-a.js'", 3])
Harness: the test ran to completion.
